Added device name to description
Added functionality to include the name of the device the volume is mounted as in the description. This doesn't _always_ match exactly, as at least in CentOS /dev/sdf is mounted as /dev/xvdf, but it at least gives you an indication as to what the drive is that is being snapshotted.